daemon: place activation code into pub module
