+++ src/lib.rs
@@ -36,6 +36,9 @@ extern crate uuid;
 #[macro_use]
 extern crate quickcheck;
 
+/// Interfaces for socket activated services.
+pub mod activation;
+
 /// Interfaces for systemd-aware daemons.
 pub mod daemon;
